Merge work/20260614-audit-consolidation into master. Lands all 18 audit
branches (Wave-1 critical correctness + Wave-2 quality sweeps). Per-branch
Codex PASS already recorded; full-solution build = 0 errors post-merge.

Conflict resolutions (both verified before merge):
- src/Invekto.Outbound/Data/ExportRepository.cs -> took consolidation's
  nullable model (int? TemplateId). Supersedes master's deployed 0-sentinel
  hotfix (3f540ab6); int? is threaded end-to-end through all Outbound
  consumers (BulkSendRepository, OutboundRepository, export param). No
  downstream breakage.
- arch/session-memory.md -> kept master HEAD (--ours); regenerated at /wrap.

Deploy: Wave-1 only in this pass (D1->D2->D3), per Q. Wave-2 (D4 quality)
follows. failclosed-auth (D3) requires live WAA Benchmark:OpsKey +
ChatAnalysis Microservice:InternalApiKey verified immediately before deploy.

+++ src/Invekto.Integrations/Data/IntegrationsRepository.cs
@@ -559,7 +559,7 @@ public sealed class IntegrationsRepository
         GetReviewRecoveryStatsAsync(int tenantId, CancellationToken ct = default)
     {
         const string sql = @"
-            SELECT recovery_status, provider, COUNT(*) as cnt
+            SELECT recovery_status, provider, COUNT(*)::int as cnt
             FROM review_alerts
             WHERE tenant_id = @tid
             GROUP BY recovery_status, provider";
